hw06/compare_lpips_ssim/requirements.txt
pandas
seaborn
numpy
matplotlib
Pillow
scikit-image
os
lpips
torch